rename kodawari.kafka to kodawari.event_streaming
