+++ py/projects/recipe_api/pyproject.toml
@@ -15,7 +15,7 @@ kodawari-authentication = {path = "../../lib/kodawari.authentication", develop =
 kodawari-models = {path = "../../lib/kodawari.models", develop = true}
 kodawari-logging-utilities = {path = "../../lib/kodawari.logging_utilities", develop = true}
 kodawari-identity = {path = "../../lib/kodawari.identity", develop = true}
-kodawari-kafka = {path = "../../lib/kodawari.kafka", develop = true}
+kodawari-event-streaming = {path = "../../lib/kodawari.event_streaming", develop = true}
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.2.0"
